requirement.txt
# --- FastAPI and Server ---
fastapi
uvicorn[standard]
gunicorn
python-dotenv
python-multipart
argparse

# --- Web UI ---
streamlit

# --- WhatsApp Integration ---
twilio

# --- AI and Language Model ---
langchain
langchain-openai
langchain-community

# --- Vector Store ---
chromadb

# --- Document Processing ---
python-docx
pypdf
docx2txt 
aiofiles
